WIP: Split off zoomodels
