+++ src/hopla/hoplalib/zoo/petmodels.py
@@ -2,9 +2,8 @@
 A helper module for Pet logic.
 """
 from dataclasses import dataclass
-from typing import Callable, Dict, Optional
+from typing import Optional
 
-from hopla.cli.groupcmds.get_user import HabiticaUser
 from hopla.hoplalib.common import GlobalConstants
 from hopla.hoplalib.errors import PrintableException
 from hopla.hoplalib.zoo.fooddata import FoodData
@@ -46,20 +45,19 @@ class Pet:
         self.pet_name = pet_name
         self.feeding_status = feeding_status
 
-        if pet_name in PetData.rare_pet_names:
-            self.__potion = None
-        else:
-            # This logic might break, but seems to be solid for past few years due to
-            # stabling naming convention by the Habitica API developers.
-            _, self.__potion = self.pet_name.split("-")
-
     def __repr__(self) -> str:
         return f"{self.__class__.__name__}({self.pet_name}: {self.feeding_status})"
 
     @property
-    def hatching_potion(self) -> str:
+    def hatching_potion(self) -> Optional[str]:
         """The hatching potion used to hatch the egg this pet came from."""
-        return self.__potion
+        if self.pet_name in PetData.rare_pet_names:
+            return None
+
+        # This logic might break, but seems to be solid for past few years due to
+        # stabling naming convention by the Habitica API developers.
+        _, potion = self.pet_name.split("-")
+        return potion
 
     def is_available(self) -> bool:
         """Return True if the feeding status says that the pet is available."""
@@ -202,94 +200,3 @@ class PetMountPair:
     def pet_available(self) -> bool:
         """True if the pet is in the pair"""
         return self.pet is not None and self.pet.is_available()
-
-
-Zoo = Dict[str, PetMountPair]
-"""
-Zoo is dictionary with key pet name keys for O(1) access to the
-PetMountPair.
-"""
-
-
-@dataclass(frozen=True)
-class ZooHelper:
-    """Class with helper functions for a Zoo."""
-    zoo: Zoo
-
-    def filter_on_pet_mount_pairs(self, predicate: Callable[[PetMountPair], bool]) -> Zoo:
-        """Filter the zoo on the pair. This does not change the underlying zoo.
-
-        :param predicate: Include the PetMountPair if the predicate returns True, else
-                          omit the pair.
-        :return: A filtered Zoo
-        """
-        return {
-            pet_name: pair for (pet_name, pair) in self.zoo.items() if predicate(pair)
-        }
-
-    def get_feedable_zoo(self) -> Zoo:
-        """Helper function to get only the pets that can be fed in this Zoo."""
-        return self.filter_on_pet_mount_pairs(PetMountPair.can_feed_pet)
-
-    def filter_on_pet_name(self, predicate: Callable[[str], bool]) -> Zoo:
-        """Filter the zoo on the pet name. This does not change the underlying zoo.
-
-        :param predicate: Include the PetMountPair if the predicate returns True, else
-                          omit the pair.
-        :return: A filtered zoo.
-        """
-        return {
-            pet_name: pair for (pet_name, pair) in self.zoo.items() if predicate(pet_name)
-        }
-
-    def filter_on_pet(self, predicate: Callable[[Pet], bool]) -> Zoo:
-        """Filter the zoo on the pet object. This does not change the underlying zoo.
-
-        :param predicate: Include the PetMountPair if the predicate returns True, else
-                          omit the pair.
-        :return: A filtered zoo.
-        """
-        return {
-            pet_name: pair for (pet_name, pair) in self.zoo.items() if predicate(pair.pet)
-        }
-
-
-class ZooBuilder:
-    """
-    Class that creates a Zoo from a HabiticaUser using the
-    builder design pattern.
-    """
-
-    def __init__(self, user: HabiticaUser):
-        self.pets: dict = user.get_pets()
-        self.mounts: dict = user.get_mounts()
-        self.__zoo: Zoo = {}  # empty until build() is called
-
-    def __repr__(self):
-        return self.__class__.__name__ + f"({self.__dict__})"
-
-    def build(self) -> Zoo:
-        """Create the Zoo. Return the Zoo in case of success"""
-
-        # loop through the pets
-        for pet_name, feed_status in self.pets.items():
-            pet = Pet(pet_name, feeding_status=FeedingStatus(feed_status))
-
-            if self.mounts.get(pet_name) is not None:
-                mount = Mount(
-                    mount_name=pet_name,
-                    availability_status=self.mounts[pet_name]
-                )
-                del self.mounts[pet_name]  # no need to handle the mount twice
-            else:
-                mount = None
-            self.__zoo[pet_name] = PetMountPair(pet=pet, mount=mount)
-
-        # loop through the remaining mounts
-        for mount_name, availability_status in self.mounts.items():
-            if mount_name not in self.pets:
-                # We found a mount without a pet. This is possible for rares.
-                mount = Mount(mount_name, availability_status=availability_status)
-                self.__zoo[mount_name] = PetMountPair(pet=None, mount=mount)
-
-        return self.__zoo
